API

In the following, you find the documentation of some, and hopefully soon all, exported functions of the NFFT.jl package family:

AbstractNFFTs.jl

AbstractNFFTs.AbstractComplexFTPlan

AbstractNFFTs.AbstractFTPlan

AbstractNFFTs.AbstractNFCTPlan

AbstractNFFTs.AbstractNFFTPlan

AbstractNFFTs.AbstractNFSTPlan

AbstractNFFTs.AbstractNNFFTPlan

AbstractNFFTs.AbstractRealFTPlan

AbstractNFFTs.accuracyParams

AbstractNFFTs.nfct

AbstractNFFTs.nfct_transpose

AbstractNFFTs.nfft

AbstractNFFTs.nfft_adjoint

AbstractNFFTs.nfst

AbstractNFFTs.nfst_transpose

AbstractNFFTs.nodes!

AbstractNFFTs.plan_nfft

AbstractNFFTs.size_in

AbstractNFFTs.size_out

NFFT.jl

NFFT.precomputeLinInterp

AbstractNFFTs.AbstractComplexFTPlan — Type

AbstractComplexFTPlan{T,D,R}

Abstract type for either an NFFT plan or an NNFFT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractFTPlan — Type

AbstractFTPlan{T,D,R}

Abstract type for any NFFT-like plan (NFFT, NFFT, NFCT, NFST).

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractNFCTPlan — Type

AbstractNFCTPlan{T,D,R}

Abstract type for an NFCT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractNFFTPlan — Type

AbstractNFFTPlan{T,D,R}

Abstract type for an NFFT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractNFSTPlan — Type

AbstractNFSTPlan{T,D,R}

Abstract type for an NFST plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractNNFFTPlan — Type

AbstractNNFFTPlan{T,D,R}

Abstract type for an NNFFT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractRealFTPlan — Type

AbstractRealFTPlan{T,D,R}

Abstract type for either an NFCT plan or an NFST plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.accuracyParams — Method

accuracyParams(; [m, σ, reltol]) -> m, σ, reltol

Calculate accuracy parameters m, σ, reltol based on either

reltol

or

m, σ

TODO: Right now, the oversampling parameter is not taken into account, i.e. σ=2.0 is assumed

AbstractNFFTs.nfct — Method

nfft(x, f, rest...; kwargs...)

calculates the nfft of the array f for the nodes contained in the matrix x The output is a vector of length M=size(nodes,2)

AbstractNFFTs.nfct_transpose — Method

nfft_adjoint(x, N, fHat, rest...; kwargs...)

calculates the adjoint nfft of the vector fHat for the nodes contained in the matrix x. The output is an array of size N

AbstractNFFTs.nfft — Method

nfft(x, f, rest...; kwargs...)

calculates the nfft of the array f for the nodes contained in the matrix x The output is a vector of length M=size(nodes,2)

AbstractNFFTs.nfft_adjoint — Method

nfft_adjoint(x, N, fHat, rest...; kwargs...)

calculates the adjoint nfft of the vector fHat for the nodes contained in the matrix x. The output is an array of size N

AbstractNFFTs.nfst — Method

nfft(x, f, rest...; kwargs...)

calculates the nfft of the array f for the nodes contained in the matrix x The output is a vector of length M=size(nodes,2)

AbstractNFFTs.nfst_transpose — Method

nfft_adjoint(x, N, fHat, rest...; kwargs...)

calculates the adjoint nfft of the vector fHat for the nodes contained in the matrix x. The output is an array of size N

AbstractNFFTs.nodes! — Method

nodes!(p, x) -> p

Change nodes x in the plan p operation and return the plan.

AbstractNFFTs.size_in — Method

size_in(p)

Size of the input array for the plan p (NFFT/NFCT/NFST/NNFFT). The returned tuple has R entries. Note that this will be the output size for the transposed / adjoint operator.

AbstractNFFTs.size_out — Method

size_out(p)

Size of the output array for the plan p (NFFT/NFCT/NFST/NNFFT). The returned tuple has R entries. Note that this will be the input size for the transposed / adjoint operator.

Base.:* — Method

    *(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the NFFT/NNFFT of a D dimensional array f of size N. fHat is a vector of length M. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

Base.:* — Method

    *(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the NFCT/NFST of a D dimensional array f of size N. fHat is a vector of length M. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

LinearAlgebra.mul! — Method

mul!(fHat, p, f) -> fHat

Inplace NFFT/NFCT/NFST/NNFFT transforming the D dimensional array f to the R dimensional array fHat. The transformation is applied along D-R+1 dimensions specified in the plan p.

LinearAlgebra.mul! — Method

mul!(f, p, fHat) -> f

Inplace adjoint NFFT/NNFFT transforming the R dimensional array fHat to the D dimensional array f. The transformation is applied along D-R+1 dimensions specified in the plan p.

LinearAlgebra.mul! — Method

mul!(f, p, fHat) -> f

Inplace transposed NFCT/NFST transforming the R dimensional array fHat to the D dimensional array f. The transformation is applied along D-R+1 dimensions specified in the plan p.

AbstractNFFTs.plan_nfft — Method

    plan_nfft(x::Matrix{T}, N::NTuple{D,Int}, rest...;  kargs...)

compute a plan for the NFFT of a size-N array at the nodes contained in x.

NFFT.precomputeLinInterp — Method

Precompute the look up table for the window function φ.

Remarks:

Only the positive half is computed

The window is computed for the interval [0, (m)/n]. The reason for the +2 is that we do evaluate the window function outside its interval, since x does not necessary match the sampling points

The window has K+1 entries and during the index calculation we multiply with the factor K/(m).

It is very important that K/(m) is an integer since our index calculation exploits this fact. We therefore always use Int(K/(m))instead of K÷(m) since this gives an error while the later variant would silently error.