API

In the following, you find the documentation of some, and hopefully soon all, exported functions of the NFFT.jl package family:

AbstractNFFTs.jl

AbstractNFFTs.AbstractNFFTPlan

AbstractNFFTs.AbstractNNFFTPlan

AbstractNFFTs.AnyNFFTPlan

AbstractNFFTs.accuracyParams

AbstractNFFTs.nfft

AbstractNFFTs.nfft_adjoint

AbstractNFFTs.nodes!

AbstractNFFTs.plan_nfft

AbstractNFFTs.size_in

AbstractNFFTs.size_out

NFFT.jl

AbstractNFFTs.AbstractNFFTPlan — Type

AbstractNFFTPlan{T,D,R}

Abstract type for an NFFT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AbstractNNFFTPlan — Type

AbstractNNFFTPlan{T,D,R}

Abstract type for an NNFFT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.AnyNFFTPlan — Type

AnyNFFTPlan{T,D,R}

Abstract type for either an NFFT plan or an NNFT plan.

T is the element type (Float32/Float64)

D is the number of dimensions of the input array.

R is the number of dimensions of the output array.

AbstractNFFTs.accuracyParams — Method

accuracyParams(; [m, σ, reltol]) -> m, σ, reltol

Calculate accuracy parameters m, σ, reltol based on either

reltol

or

m, σ

TODO: Right now, the oversampling parameter is not taken into account, i.e. σ=2.0 is assumed

AbstractNFFTs.nfft — Method

nfft(x, f::AbstractArray{T,D}, rest...; kwargs...)

calculates the NFFT of the array f for the nodes contained in the matrix x The output is a vector of length M=size(nodes,2)

AbstractNFFTs.nfft_adjoint — Method

nfft_adjoint(x, N, fHat::AbstractArray{T,D}, rest...; kwargs...)

calculates the adjoint NFFT of the vector fHat for the nodes contained in the matrix x. The output is an array of size N

AbstractNFFTs.nodes! — Method

nodes!(p, x) -> p

Change nodes x in the plan p operation and return the plan.

AbstractNFFTs.size_in — Method

size_in(p)

Size of the input array for an NFFT operation. The returned tuple has D entries. Note that this will be the output array for an adjoint NFFT.

AbstractNFFTs.size_out — Method

size_out(p)

Size of the output array for an NFFT operation. The returned tuple has R entries. Note that this will be the input array for an adjoint NFFT.

Base.:* — Method

    *(p, f) -> fHat

For a non-directional D dimensional plan p this calculates the NFFT of a D dimensional array f of size N. fHat is a vector of length M. (M and N are defined in the plan creation)

For a directional D dimensional plan p both f and fHat are D dimensional arrays, and the dimension specified in the plan creation is affected.

LinearAlgebra.mul! — Method

mul!(fHat, p, f) -> fHat

Inplace NFFT transforming the D dimensional array f to the R dimensional array fHat. The transformation is applied along D-R+1 dimensions specified in the plan p. Both f and fHat must be complex arrays of element type Complex{T}.

LinearAlgebra.mul! — Method

mul!(f, p, fHat) -> f

Inplace adjoint NFFT transforming the R dimensional array fHat to the D dimensional array f. The transformation is applied along D-R+1 dimensions specified in the plan p. Both f and fHat must be complex arrays of element type Complex{T}.

AbstractNFFTs.plan_nfft — Method

    plan_nfft(x::Matrix{T}, N::NTuple{D,Int}, rest...;  kargs...)

compute a plan for the NFFT of a size-N array at the nodes contained in x.